+++ CHANGELOG.md
@@ -1,3 +1,50 @@
+## 2.2.0+1
+
+- Compiled the example files into a single file `example.dart`.
+
+## 2.2.0
+
+- Added support for the `Crawl-delay` and `Host` fields.
+
+## 2.1.0+1
+
+- Updated README.md.
+
+## 2.1.0
+
+- Added a method `.validate()` for validating files.
+- Renamed `parser.dart` to `robots.dart`.
+
+## 2.0.1
+
+- Converted the `onlyApplicableTo` parameter in `Robots.parse()` from a `String`
+  into a `Set` to allow multiple user-agents to be specified at once.
+- Fixed the `onlyApplicableTo` parameter in `Robots.parse()` not being taken
+  into account.
+
+## 2.0.0
+
+- Additions:
+  - Added dependencies:
+    - `meta` for static analysis.
+  - Added developer dependencies:
+    - `test` for testing.
+  - Added support for the 'Sitemap' field.
+  - Added support for specifying:
+    - The precedent rule type for determining whether a certain user-agent can
+      or cannot access a certain path. (`PrecedentRuleType`)
+    - The comparison strategy to use for comparing rule precedence.
+      (`PrecedenceStrategy`)
+  - Added tests.
+- Changes:
+  - Bumped the minimum SDK version to `2.17.0` for enhanced enum support.
+- Improvements:
+  - Made all structs `const` and marked them as `@sealed` and `@immutable`.
+- Deletions:
+  - Removed dependencies:
+    - `sprint`
+    - `web_scraper`
+
 ## 1.1.1
 
 - Updated project description.